+++ src/main/scala/com/dataintuitive/viash/wrapper/BashWrapper.scala
@@ -246,6 +246,7 @@ object BashWrapper {
        |        *)  # positional arg or unknown option
        |            # since the positional args will be eval'd, can we always quote, instead of using ViashQuote
        |            VIASH_POSITIONAL_ARGS="$$VIASH_POSITIONAL_ARGS '$$1'"
+       |            [[ $$1 == -* ]] && ViashWarning $$1 looks like a parameter but is not a defined parameter and will instead be treated as a positional argument. Use "--help" to get more information on the parameters.
        |            shift # past argument
        |            ;;
        |    esac