+++ src/main/scala/com/dataintuitive/viash/functionality/Functionality.scala
@@ -274,12 +274,16 @@ case class Functionality(
   }
 
   def mainScript: Option[Script] =
-    resources.head match {
+    resources.headOption.flatMap {
       case s: Script => Some(s)
       case _ => None
     }
-
   def mainCode: Option[String] = mainScript.flatMap(_.read)
+  // provide function to use resources.tail but that allows resources to be an empty list
+  def additionalResources = resources match {
+    case _ :: tail => tail
+    case _ => List.empty[Resource]
+  }
 
   def allArgumentsAndDummies: List[Argument[_]] = allArguments ::: dummy_arguments
 }