+++ src/main/scala/com/dataintuitive/viash/platforms/NativePlatform.scala
@@ -53,7 +53,7 @@ case class NativePlatform(
     )
 
     functionality.copy(
-      resources = bashScript :: functionality.resources.tail
+      resources = bashScript :: functionality.additionalResources
     )
   }
 }