+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * `Functionality`: Add check to verify there are no double argument names or short names in the config `vsh.yaml` declarations.
 
+* `BashWrapper`: Add check to verify a parameter isn't declared twice on the CLI, except in the case `multiple: true` is declared as then it's a valid use case.
+
+* `BashWrapper`: For int min/max checking: use native bash functionality so there is no dependency to `bc`.
+  For double min/max checking: add fallback code to use `awk` in case `bc` is not present on the system (most likely to happen when running tests in a docker container).
+
 # Viash 0.5.15
 
 ## BREAKING CHANGES
@@ -73,6 +78,9 @@
         description: Description
   ```
 
+
+* Addition of the `viash_nxf_schema` component for converting a Viash config (for a workflow) into a nextflow schema file.
+
 * `NextflowVdsl3Platform`: Use `--param_list` to initialise a Nextflow channel with multiple parameter sets.
   Possible formats are csv, json, yaml, or simply a yaml_blob.
   A csv should have column names which correspond to the different arguments of this pipeline.
@@ -93,6 +101,7 @@
   taking care of the formatting in Groovy.
 
 * `NextflowVdsl3Platform`: The `--help` is auto-generated from the config.
+
 
 ## MINOR CHANGES