+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## MINOR CHANGES
 
+* `BashWrapper`: Add a warning when an argument containing flags (e.g. `--foo`) is not recognized and will be handled as a positional argument as this is likely a mistake.
+
 * `Functionality`: Add check to verify there are no double argument names or short names in the config `vsh.yaml` declarations.
 
 * `BashWrapper`: Add check to verify a parameter isn't declared twice on the CLI, except in the case `multiple: true` is declared as then it's a valid use case.