+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# Viash 0.5.16
+
+## NEW FUNCTIONALITY
+
+* `ConfigMod`: Added a `del(...)` config mod to be able to delete a value from the yaml. Example: `del(.functionality.version)`.
+
+## MINOR CHANGES
+
+* `Resources`: Handle edge case when no resources are specified in the `vsh.yaml` config file and display a warning message.
+
+* `BashWrapper`: Add a warning when an argument containing flags (e.g. `--foo`) is not recognized and will be handled as a positional argument as this is likely a mistake.
+
+* `Functionality`: Add check to verify there are no double argument names or short names in the config `vsh.yaml` declarations.
+
+* `BashWrapper`: Add check to verify a parameter isn't declared twice on the CLI, except in the case `multiple: true` is declared as then it's a valid use case.
+
+* `BashWrapper`: For int min/max checking: use native bash functionality so there is no dependency to `bc`.
+  For double min/max checking: add fallback code to use `awk` in case `bc` is not present on the system (most likely to happen when running tests in a docker container).
+
 # Viash 0.5.15
 
 ## BREAKING CHANGES
@@ -67,6 +86,9 @@
         description: Description
   ```
 
+
+* Addition of the `viash_nxf_schema` component for converting a Viash config (for a workflow) into a nextflow schema file.
+
 * `NextflowVdsl3Platform`: Use `--param_list` to initialise a Nextflow channel with multiple parameter sets.
   Possible formats are csv, json, yaml, or simply a yaml_blob.
   A csv should have column names which correspond to the different arguments of this pipeline.
@@ -87,6 +109,7 @@
   taking care of the formatting in Groovy.
 
 * `NextflowVdsl3Platform`: The `--help` is auto-generated from the config.
+
 
 ## MINOR CHANGES